Improving performance of structure generation
This changes how the clearing of blocks works so it ignores air blocks
more effeciently. instead of checking World.isAirBlock we get the block
state of the position and check if that block is air. This way we can
break mod specific air blocks and replace with vanilla air.

This check is also more effecient. If this is an air block we don't
increment the 100 block counter. This is really the magic in this since
it can go through the hundreds of blocks much faster until it finds
something it really needs to clear or it starts building the structure
proper.