Adding Saloon and other fixes
This push also fixes how paintings and item frames are generated in the world.

For paintings, this is NOT a 100% sure thing since paintings ENTIRELY depend on where on the block you place them. Because of this you have to futz with them a bit to get them in the right place.

This will be problematic when\if I do the custom pre-sets issue.