Fixing block side visibility for glass stairs/slabs
This makes it so any "glass" cubic block will cause a glass slab side to be invisible. The same is done for Glass Stairs. This will NOT be done for panes since they are shaped differently. This can work for modded glass blocks if they use the forge:glass tag.